Rebuild 1.0: You Can(not) UnitTest
What was changed?
Migrated to check library for unit tests
Using check necessitated a move to a linux based build system with
Makefiles
Restarting project to use 8b registers as a default, instead of 16b
Unit tests for 16b byte functions, get and set functions

Why was it changed?
Check seemed like a more versatile library than CuTest
The registers in the GB are 8b registers which can be considered as a
16b register in some context. The move ensures:
- we're using simple get and set operations for 8b accesses.
- the design of the emulator is more in line with GB hardware

References: