Adding actions to data model
What changed?
Added in an action column to the instruction table
Changed DBPopulator to populate the action for each operand
Updated opcodes_v view to include the action

Why?
Some instructions have an additional action taken  on one of the operands
like incrementing a register after accessing it. This was not being captured
in the previous data model.

References?