Abstracting OpcodeGenerator
What was changed?

Refactored OpcodeGenerator to be more general. Opcode case creation
now proceeds based on the mnemonic provided with the help of a
function lookup, instead of using bespoke functions for each case.

SET case function was added, still to migrate 8b LD

Why?

There was going to be a lot of common code between different operations
so trying to save as much as I can.

References?