OpcodeReport support for extended opcodes What Changed? Changed how the find_implemented opcodes function works to allow it to identify a particular function inside the cpu.c file using regex matching, and to search only that function for implemented opcodes.
The function pattern containing 'extended' now signifies that 'CB'
is to be inserted into all opcodes found in the body of this function

Why?
Because prefixed opcodes are interpreted without the prefix, there was overlap
between codes implemented in the main interpret_opcode function and the
interpret_extended_opcodes function, leading to the reports being
inaccurate.

References?