Refactor OpcodeGenerator What changed?
* Refactored OpcodeGenerator class to remove case generation logic
* Created new class heirarchy based on OpcodeClass to hold case generation logic
* Fixed a bug where connections were not closed after creating cases

Why?
* Some cases require more complicated logic than what was easily possible in OpcodeGenerator
given that some things, like the queries to run and the point at which the cycles are set
were coded in a way that meant they were not accessible to the case generation logic.

This kind of mutation was accomplished instead using object inheritance

* Bug with connections was not present before because we only maintained one with the OpcodeGenerator class.
Connections maxed out after moving to new object centered framework. Created a method to close the cursor
and DB connection, which is called after the case is created to tidy up.

References